requirements.txt
# Core dependencies for vector indexing and similarity search
faiss-cpu>=1.7.4
numpy>=1.21.0
scipy>=1.7.0

# Data manipulation and analysis
pandas>=1.3.0
scikit-learn>=1.0.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0

# Audio processing (for integration with music embeddings)
librosa>=0.9.2
soundfile>=0.10.3

# Development and testing
pytest>=6.2.4
pytest-cov>=2.12.0
black>=21.6.0
flake8>=3.9.2

# Notebook support
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and logging
pyyaml>=5.4.0
python-dotenv>=0.19.0
